Sn 34 dialogs logic

See merge request javapro9/javapro9!41

+++ src/main/java/ru/skillbox/socialnetwork/api/requests/ListUserIdsRequest.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class ListUserIdsRequest {
 
   @JsonProperty("user_ids")
-  private List<Integer> userIds;
+  private List<Long> userIds;
 }
